boichitra_api/requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.4
django-audiofield==0.10.0
django-base64field==1.0
django-celery==3.1.17
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
drf-writable-nested==0.6.3
idna==2.10
importlib-metadata==4.8.1
itypes==1.2.0
Jinja2==3.0.1
kombu==4.6.11
MarkupSafe==2.0.1
Pillow==8.2.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
zipp==3.6.0
